Item 7541: change targetedms.RunId to long (#74)

+++ lincs/src/org/labkey/lincs/LincsManager.java
@@ -179,7 +179,7 @@ public class LincsManager
         return new TableSelector(getTableInfoLincsPspJob(), new SimpleFilter(FieldKey.fromParts("Id"), id), null).getObject(LincsPspJob.class);
     }
 
-    public void deleteLincsPspJobsForRun(int runId)
+    public void deleteLincsPspJobsForRun(long runId)
     {
         Table.delete(getTableInfoLincsPspJob(), new SimpleFilter(FieldKey.fromParts("runId"), runId));
     }
